fix(index): handle dark mode
Ensure dark mode toggles on and off based on the browser theme, both by
checking that setting on the initial render and watching it for changes.
Based on the guide in https://tailwindcss.com/docs/dark-mode

Show more prominent card borders when in dark mode. Prevent the page
from overscrolling to hide inconsistencies with the background color.
Keep the footer background dark when in dark mode.